11月14日晚，CAN通信正常，DAC输出正常，开关机正常，ADC无法读取数据程序

+++ BSP/PowerModule.h
@@ -54,7 +54,7 @@
 enum{false,true};
 
 extern char ReceivedCanCommendFlag;
-extern uint16_t AdcAverage[4];
+extern uint32_t AdcAverage[4];
 extern uint16_t TargetOutputVoltage;
 extern uint16_t CurInputVoltage;
 extern uint16_t CurOutputVoltage;
